Update Patch_Helper.sh
Adding the new variable $BundelID, this now checks whether the affected application is running or not.
is executed or not. If the application is not executed, it will be executed immediately via the
jamf policy -id will be executed immediately. If it is executed or the $BundelID variable is empty,
is empty, this is included in the json and the user is informed. If all applications to be patched are not
are executed, all policies are executed in the background and the user is not aware of this.